requirements.txt
numpy==1.24.4
pandas==1.4.2
openai==1.63.2
pydantic==2.10.3
tenacity==9.1.2
loguru==0.7.3
toml==0.10.2
pillow==9.1.1
openpyxl==3.1.5
jinja2==3.1.6